+++ webextension/data/js/classes/chrome-notification-controller.js
@@ -23,7 +23,7 @@ class ChromeNotificationController {
 		this.onShownSupported = browser.notifications.hasOwnProperty('onShown');
 		if(this.onShownSupported===true){
 			browser.notifications.onShown.addListener(notificationId => {
-				ZDK.console.info(`Notification "${notificationId}" shown.`);
+				console.info(`Notification "${notificationId}" shown.`);
 				if (this.chromeNotifications.has(notificationId) && typeof this.chromeNotifications.get(notificationId).fnOnShown === "function") {
 					this.chromeNotifications.get(notificationId).fnOnShown();
 				}
@@ -43,7 +43,7 @@ class ChromeNotificationController {
 							if(typeof chromeNotifications.get(notificationId).fn === 'function'){
 								chromeNotifications.get(notificationId).fn((chromeNotifications.get(notificationId).isClosed)? 'closed' : 'timeout');
 							} else {
-								ZDK.console.warn(`${notifId} has timed out but data problem.`);
+								console.warn(`${notifId} has timed out but data problem.`);
 							}
 						}
 					});
@@ -73,7 +73,7 @@ class ChromeNotificationController {
 				const onError = (error) => {
 					if(error && typeof error.message === 'string' && (error.message === 'Adding buttons to notifications is not supported.' || error.message.indexOf("\"buttons\"") !== -1)){
 						this.chromeAPI_button_availability = false;
-						ZDK.console.debug("Buttons not supported, retrying notification without them.");
+						console.debug("Buttons not supported, retrying notification without them.");
 						if(options.buttons){
 							delete options.buttons;
 						}
@@ -125,7 +125,7 @@ class ChromeNotificationController {
 			let sound = null;
 			sendNotification(options)
 				.then(notificationId => {
-					ZDK.console.debug( `Notification "${notificationId}" created.`);
+					console.debug( `Notification "${notificationId}" created.`);
 					if (customOption!==null && typeof customOption.soundObject === 'object' && customOption.soundObject!==null && typeof customOption.soundObject.data === 'string') {
 						sound = new Audio(customOption.soundObject.data);
 						sound.volume = customOption.soundObjectVolume / 100;